Woops!
I've been committing my work thus far to my local repo. Today I realized there was a classroom project. I've mainly been focused on how to try to write it so that the CPU will always tie or win. If you comment out the call to nextMove, you can avoid the bug that mutates my whole array.. The nextMove fn is at this point intended to contain logic on how to pick a good nextMove. Right now it breaks things.